Oshirase-Neko/requirements.txt
certifi
charset-normalizer
idna
pyserial
PySocks
requests
urllib3
